# Add language support for MAXScript