--- a/grammars.yml
+++ b/grammars.yml
@@ -345,6 +345,8 @@ vendor/grammars/language-javascript:
 vendor/grammars/language-jsoniq/:
 - source.jq
 - source.xq
+vendor/grammars/language-maxscript:
+- source.maxscript
 vendor/grammars/language-ncl:
 - source.ncl
 vendor/grammars/language-python: