--- a/lib/linguist/languages.yml
+++ b/lib/linguist/languages.yml
@@ -1958,7 +1958,7 @@ M:
 
 MAXScript:
   type: programming
-  color: "#00aaaa"
+  color: "#00a6a6"
   extensions:
   - .ms
   - .mcr