This PR adds support for MAXScript, the built-in scripting language for [3ds Max](https://en.wikipedia.org/wiki/Autodesk_3ds_Max).

Usage
-----

MAXScript isn't exactly kicking JavaScript in terms of popularity, to say the least, but there's enough in-the-wild usage to warrant the language's inclusion on GitHub:

* [extension:ms fn](https://github.com/search?q=extension%3Ams+fn) (~6,610 results)
* [extension:ms function](https://github.com/search?q=extension%3Ams+function) (~13,470 results)

[ScriptSpot](http://www.scriptspot.com/) has been a well-known and popular resource for MAXScripts since at [least 2000](http://www.scriptspot.com/3ds-max/plugins/maxscribe), although the site's maintenance appears to have declined sharply in recent years. It's probably as good a time as any to introduce formal support for MAXScript on GitHub, which may encourage more MAXScripters to upload/share their work.

Heuristic modifications
-----------------------

MAXScript already offers a `move` function, so we can't rely on the current heuristic of identifying GAS by the presence of a `move` command.

My initial approach was to check for Assembly-like statements like the following, since it'd be a syntax error in MAXScript:

```
.LGCOMMAND
	.now
	.return
```

However, because of MAXScript's fluidity, it's unreasonable to assume lengthy statements won't be broken across multiple lines:
```
	value = colossal_plugin_struct_thing \
		.calculate this mumbo:#jumbo

	-- That has the same effect as writing this:
	value = colossal_plugin_struct_thing.calculate this mumbo:#jumbo
```


In other words, simply checking `\n\s*\.[A-Za-z]` isn't good enough; we need to account for things like escaped line-breaks and string literals/comments:

```rb
data.gsub(/"([^\\"]|\\.)*"|'([^\\']|\\.)*'/, "").gsub(/\\\s*(?:--.*)?\n/, "\n")
```

This might be slight overkill, but it irons out any potential for misclassification.



Performance
-----------

I took the time to analyse the performance of different heuristics when run against a [repository of considerable size](https://github.com/Alhadis/Linguist-Heuristics-MS) (15,609 files, 1.4 GiBs). I've averaged the speed results below, but the full stats can be [viewed here](http://pastebin.com/EmLwT0xU) for anybody who's curious.


### Original/Current `.ms` Heuristic

```
/((^|\s)move?[. ])|\.(include|globa?l)\s/.match(data)

Average speed (10 runs):
	real    39.8s
	user    38.9869s
	sys     0.8023s
```



### Amended Heuristic

```
/(?<!\S)\.(include|globa?l)\s/.match(data) || 
	/(?<!\/\*)(\A|\n)\s*\.[A-Za-z]/.match(data.gsub(/"([^\\"]|\\.)*"|'([^\\']|\\.)*'|\\\s*(?:--.*)?\n/, ""))

Average speed (10 runs):
	real    41.3077s
	user    40.3206s
	sys     0.9744s
```

### Earlier variations:
These are potentially slightly faster, but offer a little less accuracy:

```
/(?<!\S)\.(include|globa?l)\s/.match(data) ||
	/(?<!\/\*)(\A|\n)\s*\.[A-Za-z]/.match(data.sub(/"([^\\"]|\\.)*"|'([^\\']|\\.)*'/, ""))

Average speed (10 runs):
	real    40.0s
	user    39.1899s
	sys     0.7947s
```

```
/(?<!\S)\.(include|globa?l)\s/.match(data) || /(?<!\/\*)(\A|\n)\s*\.[A-Za-z]/.match(data)

Average speed (10 runs):
	real    38.9172s
	user    38.1532s
	sys     0.7561s
```