@larsbrinkhoff 2015-11-27
It would be nice if your updated `.ms` heuristics added some matching function for MAXScript, rather than a catch-all `else`.