@Alhadis 2015-11-27
Aye, but as I explained in the [notes for the last commit](https://github.com/Alhadis/linguist/commit/06c049b8c0533add03d2412616bef316da770d8e), real world usage of MAXScript often leaves very few distinguishing traits that can be reliably used to determine the language.

In either case, I did note your preferred method for writing heuristics (as explained at the bottom of `heuristics.rb`) but I consciously chose a catch-all approach for the reasons stated in the link above.