@Alhadis 2015-11-27
I should also mention that MAXScript runs in global scope, meaning what gets evaluated in one script is accessible to another running after it.

So, one file might look like this:
```
fn doStuff keyword:#value = (
	
)
```

... with another one looking like this:
```
doStuff 20
```

So, yeah, that's one of the reasons I took an all-inclusive approach.