@pchaigno 2015-11-29
Impressive work!

> I should also mention that MAXScript runs in global scope, meaning what gets evaluated in one script is accessible to another running after it.

But are you sure that's not the case for GAS and Groff too?
Because the heuristic rules are rarely perfect, catch-all approaches are dangerous. For those cases where none of the heuristic match, relying on the classifier (even if it's not very good) is generally preferred.

It looks like MaxScript has a number of pretty specific constructs and keywords (`--end`, `macroscript`, `local`, `#(`, `^fn [\w\s]*=`). Couldn't we use a set of those?