@Alhadis 2015-11-30
> But are you sure that's not the case for GAS and Groff too?

Groff is being evaluated first, and it's being evaluated based on standalone lines like `.TR`. Those are the commands which make up the language, and they'd be invalid declarations in MAXScript.

> Because the heuristic rules are rarely perfect, catch-all approaches are dangerous.

Well, when I ran it against the [repo](https://github.com/Alhadis/Linguist-Heuristics-MS) I mentioned, it came out with 100% accuracy for each language. These files were collected from public search results for the `.ms` file extension.

After removing duplicates (detected based on [checksum](https://github.com/Alhadis/Linguist-Heuristics-MS/blob/master/Makefile#L44)), I sorted and identified each file's language manually, assigning them to a numbered folder (1: Groff, 2: GAS, 3: MAXScript). There *was* some misclassification at first, but I refined the pattern until the results were airtight.

There were some `.ms` files that weren't GAS, Groff *or* MAXScript, but those were being misclassified anyway. I purposefully excluded them from evaluation in the aforementioned test repo; you can see [them here](https://github.com/Alhadis/Linguist-Heuristics-MS/tree/master/fixtures/public/-1).

> It looks like MaxScript has a number of pretty specific constructs and keywords

Actually, I'm afraid it doesn't:

* `--end`: `--` is the beginning of a single-line comment in MAXScript (analogous to `//` in many other languages). The `--end` is just some coder's way of identifying the ending of a particular scope, possibly for readability. I grabbed the sample files from Autodesk's tutorials, which may explain why.

* `macroscript`: Typically, MacroScripts are defined with a different file extension (`.mcr). They're rarely 
defined in `.ms` files. When 3ds Max evaluates a MacroScript clause, it generates an `.mcr` file in the user's configuration directory.

* `local`: Local (as well as global) keywords are optional modifiers that're often included more for clarity than for programmatic reasons. You can read more about the distinction [here](http://docs.autodesk.com/3DSMAX/16/ENU/MAXScript-Help/index.html?url=files/GUID-A6E95530-F419-4837-B10A-E881A62AD5F9.htm,topicNumber=d30e68307), but I won't go too far into specifics.

* `#(`: That defines an array, and not every MAXScript will have/need/use those.

* `fn [\w\s]*=)`: That's a pattern matching a `fn` expression, which may also be written `function [\w\s]*`. `fn` is just provided as a convenient shorthand (which I happen to miss when writing JavaScript).

Assembly language seems pretty well-defined based on statements that begin with `.something`, which is being checked for first *before* MAXScript. I also wrote it so any GAS files which contain `.include` or `.global` at the start of a new line are identified early as Assembly. If not, it moves on to the heuristic modifications I made/mentioned earlier - checking for statements prepended with a full-stop, as [seen here](https://github.com/Alhadis/Linguist-Heuristics-MS/blob/master/fixtures/public/2/NE.1.ms).

If it can't find any statements like those, it's quite reasonable to assume it isn't an Assembly file, or a [GT]N]roff file. And if it isn't MAXScript either, well, it was being misclassified to begin with, and the owner can use `.gitattributes` if they're unhappy with the misclassification. =)