@Alhadis 2015-12-02
@pchaigno : The heuristics are limited to affecting only the respective file extensions, correct? So having a catch-all approach for MAXScript in an `.ms` heuristic isn't going to skew Linguist's perception of Groff or GAS files with different extensions... right?