@larsbrinkhoff 2015-12-02
We should clarify that neither @pchaigno nor I are part of GitHub staff, so we can't make any decisions on this.