@Alhadis 2015-12-07
Just out of honest curiosity, who exactly gives the final go-ahead for merging the PR? Who/what are we waiting on?

Apologies if I come across impatient, I'm just uncertain as to what's happening, that's all...