@arfon 2015-12-08
Firstly @Alhadis I wanted to say thanks for submitting a really great Pull Request - I wish they were all as thorough as this :+1: 

As @pchaigno and @larsbrinkhoff mentioned, we try if possible to avoid adding 'catch all' statements to our heuristics but in this case it seems l like constructing an reliable heuristic for MAXScript is unfeasible.

I think we should get this merged in as-is.